requirements.txt
winsdk
discord-rich-presence
ytmusicapi
yt-dlp
pystray
pillow